+++ pkg/basesupplychain/baseSupplyChain_suite_test.go
@@ -1,0 +1,13 @@
+package basesupplychain_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestBaseSupplyChain(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "BaseSupplyChain Suite")
+}