+++ pkg/goboot/goboot.go
@@ -12,13 +12,14 @@ import (
 	"path/filepath"
 
 	"github.com/it-timo/goboot/pkg/baseci"
-	"github.com/it-timo/goboot/pkg/basegovernance"
 	"github.com/it-timo/goboot/pkg/basedocker"
+	"github.com/it-timo/goboot/pkg/basegovernance"
 	"github.com/it-timo/goboot/pkg/baselint"
 	"github.com/it-timo/goboot/pkg/baselocal"
 	"github.com/it-timo/goboot/pkg/baselogger"
-	"github.com/it-timo/goboot/pkg/baserelease"
 	"github.com/it-timo/goboot/pkg/baseproject"
+	"github.com/it-timo/goboot/pkg/baserelease"
+	"github.com/it-timo/goboot/pkg/basesupplychain"
 	"github.com/it-timo/goboot/pkg/basetest"
 	"github.com/it-timo/goboot/pkg/config"
 	"github.com/it-timo/goboot/pkg/goboottypes"
@@ -204,6 +205,11 @@ func (gb *GoBoot) registerMainServices() error {
 			err := gb.ServiceMgr.register(basegovernance.NewBaseGovernance(gb.cfg.TargetPath))
 			if err != nil {
 				return fmt.Errorf("failed to register %s service: %w", goboottypes.ServiceNameBaseGovernance, err)
+			}
+		case goboottypes.ServiceNameBaseSupplyChain:
+			err := gb.ServiceMgr.register(basesupplychain.NewBaseSupplyChain(gb.cfg.TargetPath))
+			if err != nil {
+				return fmt.Errorf("failed to register %s service: %w", goboottypes.ServiceNameBaseSupplyChain, err)
 			}
 		// Future services can be added here.
 		default: